backend/app/services/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Levenshtein==0.27.3
pillow==12.1.1
playwright==1.58.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyee==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-Levenshtein==0.27.3
python-multipart==0.0.22
python-telegram-bot==22.6
python-whois==0.9.6
pyzbar==0.1.9
RapidFuzz==3.14.3
redis==7.2.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
